@@ -1,15 +1,15 @@
 buildscript {
-    ext.kotlin_version = '1.4.21'
+    ext.kotlin_version = '1.5.10'
     repositories {
         google()
-        jcenter()
+        mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.1.1'
+        classpath 'com.android.tools.build:gradle:4.2.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
 
-        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:2.3.2"
-        classpath 'com.google.gms:google-services:4.3.4'
+        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:2.3.5"
+        classpath 'com.google.gms:google-services:4.3.8'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -18,12 +18,13 @@ buildscript {
 
 plugins {
     id("io.gitlab.arturbosch.detekt").version("1.15.0-RC1")
+    id("org.jetbrains.kotlin.android").version("1.4.20").apply(false)
 }
 
 allprojects {
     repositories {
         google()
-        jcenter()
+        mavenCentral()
     }
 }
 
