@@ -3,13 +3,14 @@ package net.it96.enfoque
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
+import android.util.Log
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.firebase.ui.auth.AuthUI
+import com.firebase.ui.auth.ErrorCodes
 import com.firebase.ui.auth.IdpResponse
 import com.google.firebase.auth.FirebaseAuth
-import kotlinx.android.synthetic.main.activity_login.*
-import timber.log.Timber
+import net.it96.enfoque.databinding.ActivityLoginBinding
 
 
 class LoginActivity : AppCompatActivity() {
@@ -18,9 +19,15 @@ class LoginActivity : AppCompatActivity() {
         private const val RC_SIGN_IN = 423
     }
 
+    private val TAG = "LoginActivity"
+    private var _binding: ActivityLoginBinding? = null
+    private val binding get() = _binding!!
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_login)
+        _binding = ActivityLoginBinding.inflate(layoutInflater)
+        val view = binding.root
+        setContentView(view)
         login()
     }
 
@@ -33,7 +40,7 @@ class LoginActivity : AppCompatActivity() {
 //            AuthUI.IdpConfig.FacebookBuilder().build(),
 //            AuthUI.IdpConfig.TwitterBuilder().build()
         )
-        btn_login.setOnClickListener {
+        binding.btnLogin.setOnClickListener {
             startActivityForResult(
                 AuthUI.getInstance()
                     .createSignInIntentBuilder()
@@ -46,22 +53,30 @@ class LoginActivity : AppCompatActivity() {
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         super.onActivityResult(requestCode, resultCode, data)
-
         if (requestCode == RC_SIGN_IN) {
             val response = IdpResponse.fromResultIntent(data)
-
             if (resultCode == Activity.RESULT_OK) {
                 // Successfully signed in
                 val user = FirebaseAuth.getInstance().currentUser
-                Timber.i("***MZP*** user: ${user?.email}")
+                Log.i(TAG, "***MZP*** user: ${user?.email}")
                 startActivity(Intent(this, ProjectActivity::class.java))
                 finish()
             } else {
+//                if(response == null) {
+//                    finish()
+//                }
                 // Sign in failed. If response is null the user canceled the
                 // sign-in flow using the back button. Otherwise check
                 // response.getError().getErrorCode() and handle the error.
                 // ...
-                Toast.makeText(this, "Ocurrio un error ${response!!.error!!.message}", Toast.LENGTH_LONG).show()
+                if(response?.error?.errorCode == ErrorCodes.DEVELOPER_ERROR) {
+                    Toast.makeText(this,
+                        "Ocurrio un error ${response.error!!.message.toString()}",
+                        Toast.LENGTH_LONG).show()
+                    Toast.makeText(this,
+                        "response.getError()?.getErrorCode() ${response.getError()?.getErrorCode()}",
+                        Toast.LENGTH_LONG).show()
+                }
             }
         }
     }
